Move all metrics helpers to separate file.
